requirements.txt
google-genai
streamlit
requests
python-dotenv
pandas
plotly
pillow